requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-socketio[asgi]==5.11.3
pydantic==2.9.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.36
aiofiles==24.1.0
python-multipart==0.0.9
jinja2==3.1.4
